added UserDetailsImpl and some changes

+++ src/main/java/ua/foxminded/javaspring/kocherga/web_application/models/UserRole.java
@@ -28,7 +28,7 @@ public class UserRole {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(long userId) {
         this.userId = userId;
     }
 
@@ -36,7 +36,7 @@ public class UserRole {
         return roleId;
     }
 
-    public void setRoleId(int roleId) {
+    public void setRoleId(long roleId) {
         this.roleId = roleId;
     }
